refactor: Send payload from child component to parent component
