+++ src/modules/Article/components/Article/FormArticleDetail.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { toRaw } from "vue"
+import { toRaw, ref } from "vue"
 import useFormArticleDetail from "../../composables/Article/useFormArticleDetail";
 // import type { RadioOption } from "@/types/RadioOption";
 import type { ArticleDetail } from "../../types/Article/ArticleDetail";
@@ -17,22 +17,35 @@ const {
   v$
 } = useFormArticleDetail(props.article_detail)
 
+const selectedPresentations = ref([])
+
 const emits = defineEmits<{
   (e: 'submitArticleDetail', form: ArticleDetail): void
 }>()
 
 const submitArticleDetail = async () => {
-  const result = await v$.value.$validate();
+  alert(selectedPresentations.value)
+  /*const result = await v$.value.$validate();
   if (result) {
     emits("submitArticleDetail", toRaw(form));
-  }
+  }*/
+}
+
+const selectPresentationId = (presentationId) => {
+  selectedPresentations.value.includes(presentationId)
+    ? selectedPresentations.value.splice(selectedPresentations.value.indexOf(presentationId),1)
+      : selectedPresentations.value.push(presentationId)
 }
 </script>
 
 <template>
   <div>
+    <TableSearchArticleDetail
+      :selectedPresentations="selectedPresentations"
+       @selectPresentationId="selectPresentationId"
+    /> 
     <form @submit.prevent="submitArticleDetail">       
-      <TableSearchArticleDetail />      
+     
       <div class="block flex justify-center mt-4">
         <AppBtn
           type="submit"
